# test_runner/performance: add logical message ingest benchmark