--- a/test_runner/performance/test_ingest_logical_message.py
+++ b/test_runner/performance/test_ingest_logical_message.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-from contextlib import closing
-
 import pytest
 from fixtures.benchmark_fixture import MetricReport, NeonBenchmarker
 from fixtures.common_types import Lsn
@@ -16,7 +14,7 @@ from fixtures.pg_version import PgVersion
 
 @pytest.mark.timeout(600)
 @pytest.mark.parametrize("size", [1024, 8192, 131072])
-@pytest.mark.parametrize("fsync", [True, False])
+@pytest.mark.parametrize("fsync", [True, False], ids=["fsync", "nofsync"])
 def test_ingest_logical_message(
     request: pytest.FixtureRequest,
     neon_env_builder: NeonEnvBuilder,
@@ -40,31 +38,30 @@ def test_ingest_logical_message(
     # Ingest data and measure durations.
     start_lsn = Lsn(endpoint.safe_psql("select pg_current_wal_lsn()")[0][0])
 
-    with closing(endpoint.connect()) as conn:
-        with conn.cursor() as cur:
-            cur.execute("set statement_timeout = 0")
+    with endpoint.cursor() as cur:
+        cur.execute("set statement_timeout = 0")
 
-            # Postgres will return once the logical messages have been written to its local WAL,
-            # without waiting for Safekeeper commit. We measure ingestion time both for Postgres,
-            # Safekeeper, and Pageserver to detect bottlenecks.
-            with zenbenchmark.record_duration("pageserver_ingest"):
-                with zenbenchmark.record_duration("safekeeper_ingest"):
-                    with zenbenchmark.record_duration("postgres_ingest"):
-                        log.info("Ingesting data")
-                        cur.execute(f"""
-                            select pg_logical_emit_message(false, '', repeat('x', {size}))
-                            from generate_series(1, {count})
-                        """)
+        # Postgres will return once the logical messages have been written to its local WAL, without
+        # waiting for Safekeeper commit. We measure ingestion time both for Postgres, Safekeeper,
+        # and Pageserver to detect bottlenecks.
+        log.info("Ingesting data")
+        with zenbenchmark.record_duration("pageserver_ingest"):
+            with zenbenchmark.record_duration("safekeeper_ingest"):
+                with zenbenchmark.record_duration("postgres_ingest"):
+                    cur.execute(f"""
+                        select pg_logical_emit_message(false, '', repeat('x', {size}))
+                        from generate_series(1, {count})
+                    """)
 
-                        end_lsn = Lsn(endpoint.safe_psql("select pg_current_wal_lsn()")[0][0])
+                    end_lsn = Lsn(endpoint.safe_psql("select pg_current_wal_lsn()")[0][0])
 
-                    # Wait for Safekeeper.
-                    log.info("Waiting for Safekeeper to catch up")
-                    wait_for_commit_lsn(env, env.initial_tenant, env.initial_timeline, end_lsn)
+                # Wait for Safekeeper.
+                log.info("Waiting for Safekeeper to catch up")
+                wait_for_commit_lsn(env, env.initial_tenant, env.initial_timeline, end_lsn)
 
-                # Wait for Pageserver ingestion.
-                log.info("Waiting for Pageserver to catch up")
-                wait_for_last_flush_lsn(env, endpoint, env.initial_tenant, env.initial_timeline)
+            # Wait for Pageserver.
+            log.info("Waiting for Pageserver to catch up")
+            wait_for_last_flush_lsn(env, endpoint, env.initial_tenant, env.initial_timeline)
 
     # Now that all data is ingested, delete and recreate the tenant in the pageserver. This will
     # reingest all the WAL from the safekeeper without any other constraints. This gives us a