--- a/test_runner/performance/test_ingest_logical_message.py
+++ b/test_runner/performance/test_ingest_logical_message.py
@@ -10,7 +10,6 @@ from fixtures.neon_fixtures import (
     wait_for_last_flush_lsn,
 )
 from fixtures.pageserver.utils import wait_for_last_record_lsn
-from fixtures.pg_version import PgVersion
 
 
 @pytest.mark.timeout(600)
@@ -80,9 +79,6 @@ def test_ingest_logical_message(
     # Now that all data is ingested, delete and recreate the tenant in the pageserver. This will
     # reingest all the WAL from the safekeeper without any other constraints. This gives us a
     # baseline of how fast the pageserver can ingest this WAL in isolation.
-    pg_version = PgVersion(
-        client.timeline_detail(env.initial_tenant, env.initial_timeline)["pg_version"]
-    )
     status = env.storage_controller.inspect(tenant_shard_id=env.initial_tenant)
     assert status is not None
 
@@ -91,7 +87,7 @@ def test_ingest_logical_message(
 
     with zenbenchmark.record_duration("pageserver_recover_ingest"):
         log.info("Recovering WAL into pageserver")
-        client.timeline_create(pg_version, env.initial_tenant, env.initial_timeline)
+        client.timeline_create(env.pg_version, env.initial_tenant, env.initial_timeline)
         wait_for_last_flush_lsn(env, endpoint, env.initial_tenant, env.initial_timeline)
 
     # Emit metrics.