@erikgrinaker 2024-11-14
> Would you mind noting the results you currently get in the PR description? I find this useful when looking back at history of benchmarks (even though we also have some history in a database)

Sure, added a result from my MacBook, but I don't think it's particularly useful without more context and representative hardware.

For running analysis, see https://github.com/neondatabase/neon/issues/9642#issuecomment-2475995205.